+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## Added
 - multisite support
 - stub for onlyoffice-wordpress-block that displays error information if the document server is not accessible or the file is not found 
-- settings for onlyoffice block (width, height, align, show open in onlyoffice bytton)
+- settings for onlyoffice block (width, height, align, show open in onlyoffice button)
 - button copy link to editor, in the table with files
 - ability to insert a link to the editor on the page
 - access for an anonymous user, open the document for viewing in the editor if the document is attached to a public post